% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_select_script}
\alias{add_select_script}
\title{Selecting for Leaflet}
\usage{
add_select_script(lf, styleFalse, styleTrue, ns = "")
}
\description{
Taken from the interals of the \link{mapedit} package
}
\keyword{internal}
