% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base_map}
\alias{base_map}
\title{Base Map}
\usage{
base_map()
}
\value{
A leaflet map with provider layers: \strong{"Esri.WorldImagery", "CartoDB.Positron",
"OpenStreetMap", "CartoDB.DarkMatter", "OpenTopoMap"
"Hydrography"}
}
\description{
A generic leaflet base map used in the shiny apps.
}
