% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{basinModUI}
\alias{basinModUI}
\title{Shiny Module UI for basins}
\usage{
basinModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
UI function for Shiny module
}
\description{
A shiny Module to.
}
