% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_sf_geocollection}
\alias{convert_sf_geocollection}
\title{Convert GEOMETRYCOLLECTION to POLYGONS}
\usage{
convert_sf_geocollection(x)
}
\arguments{
\item{x}{A sf object}
}
\value{
A converted sf object from GEOMETRYCOLLECTION to POLYGON or MULTIPOLYGON.
}
\description{
Convert GEOMETRYCOLLECTION to POLYGONS
}
