% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_conditions}
\alias{current_conditions}
\title{Get Current Conditions}
\usage{
current_conditions()
}
\value{
a \code{tibble} with current conditions and attributes from USGS dashboard.
}
\description{
Get Current Conditions
}
\note{
The time zone used in the URL call is the R session time zone. Also, the time is 1-hour behind.
Here are the attributes that are with the data.frame: AgencyCode,SiteNumber,SiteName,SiteTypeCode,Latitude,Longitude,
CurrentConditionID,ParameterCode,TimeLocal,TimeZoneCode,Value,
ValueFlagCode,RateOfChangeUnitPerHour,StatisticStatusCode,FloodStageStatusCode.
}
