% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_site_new}
\alias{df_site_new}
\title{Retired USGS Sites}
\format{
\subsection{\code{df_site_new}}{

A data frame with 14,966 rows and 4 columns:
\describe{
\item{SiteNumber}{Site Number}
\item{SiteName}{Site Name}
\item{location.coordinates1}{Longitude coordinates}
\item{location.coordinates2}{Latitude coordinates}
...
}
}
}
\usage{
df_site_new
}
\description{
A subset of data that has retired USGS sites and coordinates.
}
\keyword{datasets}
