% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Basin}
\alias{get_Basin}
\title{Get Basin Boundary NLDI}
\usage{
get_Basin(point)
}
\arguments{
\item{point}{A sf point object.}
}
\value{
An sf object with added \code{comid} and \code{basin}.
}
\description{
This function uses the USGS water data API to link a point to a realized basin. This is
not the same as delineating from the exact point, rather this API uses NLDI to find the closest
basin downstream source point. There is a lot you can do with this API and I would recommend
looking at {nhdplusTools} as that has a lot of functionality and better documentation.
}
\note{
\code{point} needs geometry column.
}
