% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_NLDI}
\alias{get_NLDI}
\title{Get NLDI}
\usage{
get_NLDI(point)
}
\arguments{
\item{point}{A sf point.}
}
\value{
A list of UT, UM and basin boundary sf objects
}
\description{
This function grabs the upstream tributaries, upstream main stream and basin boundary using
the NLDI API. It then combines the NLDI zonal stats to the basin boundary shape, i.e. 'TOT' is the 'total' basin zonal statistic.
}
