% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_NLDI_catchments}
\alias{get_NLDI_catchments}
\title{Get NLDI Catchments}
\usage{
get_NLDI_catchments(point, type = "local", method = "all")
}
\arguments{
\item{point}{A sf point object.}

\item{type}{A \code{character} 'local' or 'all'.}

\item{method}{A \code{character} 'local' or 'all'.}
}
\value{
A list of sf objects: UT and catchments.
}
\description{
This function grabs the 'local' zonal stats for 'all' subcatchments above a point or only for
the 'local' catchment using the NLDI API. This is different than get_NLDI(), which grabs the entire basin above a point.
}
\note{
This function can be expensive when using type = 'local' and method = 'all' depending on
the size of the upstream area.
}
