% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_noaatlas}
\alias{get_noaatlas}
\title{Get National Oceanic and Atmospheric Administration (NOAA) Atlas 14}
\usage{
get_noaatlas(point, data_type = "depth", units = "english", series = "pds")
}
\arguments{
\item{point}{An sf POINT object.}

\item{data_type}{A character, e.g. 'depth' or 'intensity'.}

\item{units}{A character, e.g. 'english' or 'metric'.}

\item{series}{A character, e.g. 'pds' or 'ams'}
}
\value{
A tibble with quantiles, lower and upper bounds.
}
\description{
Get National Oceanic and Atmospheric Administration (NOAA) Atlas 14
}
