% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_noaatlas_png}
\alias{get_noaatlas_png}
\title{Get National Oceanic and Atmospheric Administration (NOAA) Atlas 14 Graphics}
\usage{
get_noaatlas_png(
  point,
  data_type = "depth",
  units = "english",
  series = "pds",
  dur,
  print = TRUE,
  destfile = NULL
)
}
\arguments{
\item{point}{An sf POINT object.}

\item{data_type}{A character, e.g. 'depth' or 'intensity'.}

\item{units}{A character, e.g. 'english' or 'metric'.}

\item{series}{A character, e.g. 'pds' or 'ams'}

\item{dur}{A character, e.g. '10m', '4d', '24h', etc.}

\item{print}{Logical, default printing of PNG in Rstudio Viewer}

\item{destfile}{A character file path, default in NULL.}
}
\value{
A Portable Network Graphic printed to Rstudio Viewer.
}
\description{
Get National Oceanic and Atmospheric Administration (NOAA) Atlas 14 Graphics
}
\note{
If \code{print = FALSE}, then the Graphic will be saved to tempfile or destination.
}
