% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_whitebox_basin}
\alias{get_whitebox_basin}
\title{whitebox helpers}
\usage{
get_whitebox_basin(
  sf_point,
  z,
  snap_dist,
  smoothing = TRUE,
  depressions = TRUE,
  ...
)
}
\arguments{
\item{sf_point}{a sf data.frame point(s)}

\item{z}{param for elevatr function get_elev_raster()}

\item{snap_dist}{distance to snap to stream (in meters)}

\item{smoothing}{logical}

\item{depressions}{logical}

\item{...}{arguments to pass to whitebox tools functions}
}
\value{
a sf polygon
}
\description{
whitebox helpers
}
