% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_whitebox_streams}
\alias{get_whitebox_streams}
\title{whitebox helpers for streams}
\usage{
get_whitebox_streams(aoi, z, threshold, prj)
}
\arguments{
\item{aoi}{a sf polygon}

\item{z}{param for elevatr function get_elev_raster()}

\item{threshold}{numeric; cell threshold for stream delineation}

\item{prj}{A character vector with proj4string.}
}
\value{
a list of file paths to .tif files and a terra::rast object
}
\description{
whitebox helpers for streams
}
