% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nldi_basin_function}
\alias{nldi_basin_function}
\title{Calling NLDI API}
\usage{
nldi_basin_function(point)
}
\arguments{
\item{point}{sf data.frame}
}
\value{
a sf data.frame with watershed basin
}
\description{
Calling NLDI API
}
