% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{noaatlasMod}
\alias{noaatlasMod}
\title{Shiny Module Server for National Oceanic and Atmospheric Administration (NOAA) Atlas 14}
\usage{
noaatlasMod(input, output, session, values, map)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{values}{A reactive Values list to pass}

\item{map}{A background leaflet or mapview map to be used for editing. If NULL a blank mapview canvas will be provided.}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for National Oceanic and Atmospheric Administration (NOAA) Atlas 14
}
