% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{noaatlasModUI}
\alias{noaatlasModUI}
\title{Shiny Module UI for National Oceanic and Atmospheric Administration (NOAA) Atlas 14}
\usage{
noaatlasModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
UI function for Shiny module
}
\description{
A shiny Module to get numerous precipitation-frequency statistics.
}
