% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{streamnetworkMod}
\alias{streamnetworkMod}
\title{Shiny Module Server for stream networks}
\usage{
streamnetworkMod(input, output, session, values, dem, threshold = 1000, map)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{values}{A reactive Values list to pass}

\item{dem}{A raster or terra object dem. (optional)}

\item{threshold}{A threshold for stream initiation. 1000 (default).}

\item{map}{a background leaflet or mapview map to be used for editing. If NULL a blank mapview canvas will be provided.}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for stream networks
}
