% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{streamnetworkModUI}
\alias{streamnetworkModUI}
\title{Shiny Module UI for stream network generation}
\usage{
streamnetworkModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
UI function for Shiny module
}
\description{
A shiny Module to.
}
