% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{usgsdvMod}
\alias{usgsdvMod}
\title{Shiny Module Server for United States Geologic Survey (USGS) daily values}
\usage{
usgsdvMod(input, output, session, values)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{values}{A reactive Values list to pass}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for United States Geologic Survey (USGS) daily values
}
