% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{usgsModUI}
\alias{usgsModUI}
\title{Shiny Module UI for USGS instantaneous values}
\usage{
usgsModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
UI function for Shiny module
}
\description{
A shiny Module to.
}
