% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{basinMod}
\alias{basinMod}
\title{Shiny Module Server for basin generation}
\usage{
basinMod(input, output, session, values, dem, map)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{values}{A reactive Values list to pass}

\item{dem}{A raster or terra object dem.}

\item{map}{a background leaflet or mapview map to be used for editing. If NULL a blank mapview canvas will be provided.}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for basin generation
}
