% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.R
\name{get_basin_interactively}
\alias{get_basin_interactively}
\title{Get Watershed Basin Interactively}
\usage{
get_basin_interactively(
  ns = "basin-ui",
  viewer = shiny::paneViewer(),
  title = "Delineate Basin",
  ...
)
}
\arguments{
\item{ns}{\code{string} name for the Shiny \code{namespace} to use.  The \code{ns}
is unlikely to require a change.}

\item{viewer}{\code{function} for the viewer.  See Shiny \code{\link[shiny]{viewer}}.
NOTE: when using \code{browserViewer(browser = getOption("browser"))} to
open the app in the default browser, the browser window will automatically
close when closing the app (by pressing "done" or "cancel") in most browsers.
Firefox is an exception. See Details for instructions on how to enable this
behaviour in Firefox.}

\item{title}{\code{string} to customize the title of the UI window.  The default
is "Delineate Basin".}

\item{...}{other arguments to \code{leafletOutput()} in module.}
}
\value{
A list of sf objects that contain watershed polygons
the user collected during shiny session.
}
\description{
This function allows the user to delineate watershed basins interactively with a
shiny app. It uses the {elevatr} package to acquire the Digital Elevation Model (DEM)
and {whitebox} package to delineate the basin (see details).
}
\details{
\strong{This function will throw an error if you don't draw the bounding box (rectangle) first.}
Once the user has drawn the bounding box then you can use the marker as a pour point location.

\strong{Steps}
\enumerate{
\item Input a well-suited DEM zoom level, stream threshold (resolution) and snapping distance.
\item Draw bounding box (rectangle or polygon).
\item Use marker to place pour point(s).
\item If necessary, change 'Cell Threshold' to change drainage density.
\item Repeat steps 1-4 if needed.
\item When finished, press 'done' and basins will be saved as a list in local environment.
}

In addition, this function uses both \code{whitebox::wbt_feature_preserving_smoothing()} and \code{whitebox::wbt_breach_depressions()}
prior to running the flow direction and flow accumulation (both d8) algorithms.
}
\note{
The marker will only work for the most current stream raster. You can have multiple areas but
you need to make sure that you are on the most current raster or the app will crash.
}
\examples{

if(interactive()){
basin_data <- get_basin_interactively()
}


}
