% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus.R
\name{get_nhdplus_interactively}
\alias{get_nhdplus_interactively}
\title{Get NHDPlus Interactively}
\usage{
get_nhdplus_interactively()
}
\value{
A list of sf objects that the user collected during shiny session.
}
\description{
This function allows the user go get NHDPlus realizations interactively with a
shiny app.
}
\examples{
\dontrun{


nhdplus_data <- get_nhdplus_interactively()



}
#' @importFrom shiny NS tagList
}
