% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamnetwork.R
\name{get_stream_network_interactively}
\alias{get_stream_network_interactively}
\title{Get Stream Network Interactively}
\usage{
get_stream_network_interactively(
  ns = "basin-ui",
  viewer = shiny::paneViewer(),
  title = "Delineate Basin",
  ...
)
}
\arguments{
\item{ns}{\code{string} name for the Shiny \code{namespace} to use.  The \code{ns}
is unlikely to require a change.}

\item{viewer}{\code{function} for the viewer.  See Shiny \code{\link[shiny]{viewer}}.
NOTE: when using \code{browserViewer(browser = getOption("browser"))} to
open the app in the default browser, the browser window will automatically
close when closing the app (by pressing "done" or "cancel") in most browsers.
Firefox is an exception. See Details for instructions on how to enable this
behaviour in Firefox.}

\item{title}{\code{string} to customize the title of the UI window.  The default
is "Delineate Basin".}

\item{...}{other arguments to \code{leafletOutput()} in module.}
}
\value{
A list of sf objects that the user collected during shiny session.
}
\description{
This function allows the user to get stream networks and watersheds interactively with a
shiny app. It uses the {elevatr} package to acquire the Digital Elevation Model (DEM)
and {whitebox} package to delineate the stream network and watersheds (see details).
}
\details{
This function uses the package \link{elevatr} to download the DEM
Once the user has drawn the bounding box then you can use the marker as a pour point location.

\strong{Steps}
\enumerate{
\item Input a well-suited DEM zoom level, stream threshold (cells) and snapping distance.
\item Draw bounding box (rectangle).
\item Wait for layers to respond.
\item when finished, press 'done' and stream network and watersheds will be saved as a list in local environment.
}

In addition, this function uses both \code{whitebox::wbt_feature_preserving_smoothing()} and \code{whitebox::wbt_breach_depressions()}
prior to running the flow direction and flow accumulation (both d8) algorithms.
}
\examples{

if(interactive()){
nhdplus_data <- get_stream_network_interactively()
}



}
