% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_whitebox_streams}
\alias{get_whitebox_streams}
\title{whitebox helpers for streams}
\usage{
get_whitebox_streams(
  aoi,
  z,
  smoothing = TRUE,
  depressions = TRUE,
  threshold,
  ...
)
}
\arguments{
\item{aoi}{a sf polygon}

\item{z}{param for elevatr function get_elev_raster()}

\item{smoothing}{logical}

\item{depressions}{logical}

\item{threshold}{numeric; cell threshold for stream delineation}

\item{...}{arguments to pass to whitebox tools functions}
}
\value{
a list of file paths to .tif files and a terra::rast object
}
\description{
whitebox helpers for streams
}
