% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_whitebox_ws}
\alias{get_whitebox_ws}
\title{whitbox helpers for watersheds}
\usage{
get_whitebox_ws(sf_point, output_streams, output_pointer, snap_dist)
}
\arguments{
\item{snap_dist}{}
}
\value{
A sf polygon of a watershed
}
\description{
whitbox helpers for watersheds
}
