% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename Geometry Column}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{A sf object}

\item{name}{character.}
}
\value{
A sf object with a renamed geometry column.
}
\description{
Rename Geometry Column
}
\note{
This function was grabbed from \href{https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r}{stack overflow} from the legend spacedman.
}
